requirements.txt
surya-ocr
Pillow
python-doctr
easyocr
termcolor
pytesseract
openai
matplotlib
google.generativeai

